added updated version event to resolver

+++ src/interfaces/resolvers/IVersionableResolver.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.17;
 // resolver interface based on ENS
 // https://github.com/ensdomains/ens-contracts
 interface IVersionableResolver {
-    event VersionChanged(bytes32 indexed node, uint64 newVersion);
+    event VersionChanged(bytes32 indexed node, uint256 newVersion);
 
     //cheaper gas using uint256
     function recordVersions(bytes32 node) external view returns (uint256);
